Implement beginnings of ASCII character rendering functionality
The ASCIITileset class keeps a record of a texture atlas, created from a texture containing ASCII glyphs, from top-left to bottom-right standard ASCII order. It has methods to retrieve the corresponding Texture2dRegions for each glyph.

The ASCIIGrid class for now only handles the drawing of glyphs with foreground and background color. In future it may be benifical to implement arrays to act as a kind of buffer from which to update the glyphs and colours, then draw to the screen separately.